requirements.txt
streamlit
langchain
langgraph
langchain-google-genai
beautifulsoup4
python-dotenv
requests
tavily-python
faiss-cpu              
langchain-community